requirements.txt
Django==2.2.7
djangorestframework==3.9.2
psycopg2-binary==2.8.4
django-extensions==2.2.5
django-filter==2.2.0